# Improve quick open performance and behaviour