This commit adds a one time cache step to store filenames and icon references. Previously, the file system would be scanned every search update. The diff is somewhat larger than it really is, I moved some methods to a better position.

**Behaviour**
 - Always scroll back to top (selection) upon refresh.
 - Scoring function: fix second case and remove expensive similarity computation for insignificant results.
 - Only show a max amount of files (currently set at 300)

**Performance**
-  Cache filepaths and icons at popup.
-  Sort files with heapsort instead of selection sort.
-  Fix double search by clearing the text

<details><summary>Show metrics</summary>

Some rough metrics for `_update_search` in a project with 500 scripts and 400 scenes in Quick Open (not scene/script). Compiled with `debug`, no debugger attached.
| Query | Master (ms) | PR (ms) |
| --- | --:  | --: |
''    | 11,3 | 3,3
'1'   | 15,7 | 4,9
'12'  | 5,5  | 1,6
'123' | 3,2  | 0,9
'1234'| 4,2  | 0,6

These results are without the one time cost of building the cache (~7ms for this amount of files). 
Project: [500Custom.zip](https://github.com/godotengine/godot/files/4918762/500Custom.zip)
</details>

Some comments:
- The dialog supports selecting multiple files, but this adds complexity and is unusual for QuickOpen. Could be removed.
- Clearing the search box triggers an update, which makes the popup code somewhat ugly. Could remove option to not clear.